refactor(crestodian): extract setup auth options
